backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
requests
pydantic
motor
